# admin/src/main/resources/public/swagger/index.yaml
swagger: '2.0'
info:
  description: 是一套轻量级权限管理系统，极低门槛，拿来即用。设计之初，就非常注重安全性，为企业系统保驾护航，让一切都变得如此简单。
  version: 3.0.0
  title: renren-security
basePath: /renren-security
schemes:
  - http

#定义接口数据
paths:
...
  /sys/dept/info/{deptId}:
    get:
      tags:
        - 社区管理
      summary: 获取社区信息
      produces:
        - application/json
      parameters:
        - name: deptId
          description: 社区ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回社区信息
          schema:
            $ref: '#/definitions/SysDeptEntity'
  /sys/dept/save:
    post:
      tags:
        - 社区管理
      summary: 添加社区
      produces:
        - application/json
      parameters:
        - name: body
          description: 社区对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysDeptEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dept/update:
    post:
      tags:
        - 社区管理
      summary: 修改社区
      produces:
        - application/json
      parameters:
        - name: body
          description: 社区对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysDeptEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dept/delete:
    post:
      tags:
        - 社区管理
      summary: 删除社区
      produces:
        - application/json
      parameters:
        - name: body
          description: 社区ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  SysDeptEntity:
    type: object
    properties:
      deptId:
        description: 社区ID
        type: integer
        format: int64
      name:
        description: 社区名称
        type: string
      parentId:
        description: 上级社区ID，一级社区为0
        type: integer
        format: int64
      parentName:
        description: 上级社区名称
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
      open:
        description: 是否展开   true：展开   false：不展开
        type: boolean
        format: int32
  SysDeptEntityEdit:
    type: object
    properties:
      name:
        description: 社区名称
        type: string
      parentId:
        description: 上级社区ID，一级社区为0
        type: integer
        format: int64
      orderNum:
        description: 排序
        type: integer
        format: int32
...